move SchemaId from persist-client to persist-types
