+++ src/storage-client/src/controller.rs
@@ -32,8 +32,8 @@ use mz_cluster_client::client::ClusterReplicaLocation;
 use mz_cluster_client::ReplicaId;
 use mz_ore::collections::CollectionExt;
 use mz_persist_client::read::{Cursor, ReadHandle};
-use mz_persist_client::schema::SchemaId;
 use mz_persist_client::stats::{SnapshotPartsStats, SnapshotStats};
+use mz_persist_types::schema::SchemaId;
 use mz_persist_types::{Codec64, Opaque, ShardId};
 use mz_repr::{Diff, GlobalId, RelationDesc, Row};
 use mz_storage_types::configuration::StorageConfiguration;
